+++ src/anonymizer.js
@@ -112,7 +112,10 @@ function(
 			startXSave: 0,
 			startYSave: 0,
 			startX: 0,
-			startY: 0,
+			startY: 0
+		};
+
+		this._flags = {
 			wasRightClick: false,
 			wasMove: false,
 			wasPreview: false,
@@ -586,13 +589,13 @@ function(
 		this._mouse.startYSave = e.offsetY;
 		this._mouse.startX = e.offsetX;
 		this._mouse.startY = e.offsetY;
-		this._mouse.wasMove = false;
-		this._mouse.wasRightClick = this._isRightClick(e);
+		this._flags.wasMove = false;
+		this._flags.wasRightClick = this._isRightClick(e);
 
 		// circle
 		if (this._opts.curEntity == $anonymizer.ENTITES.CIRCLE) {
-			this._mouse.wasImgMove = false;
-			this._mouse.wasPreview = false;
+			this._flags.wasImgMove = false;
+			this._flags.wasPreview = false;
 
 			this._canvas.addEventListener("mousemove", this._binds.mouseMove);
 			this._canvas.addEventListener("mouseup", this._binds.mouseUp);
@@ -606,15 +609,15 @@ function(
 			lineCanvas.height = this._opts.canHeight;
 			lineCanvas.classList.add("line-canvas");
 
-			this._mouse.wasPreview = false;
-			this._mouse.wasLine = false;
+			this._flags.wasPreview = false;
+			this._flags.wasLine = false;
 
 			this._lineCanvas = lineCanvas;
 			this._lineCanvas.addEventListener("mousemove", this._binds.mouseMoveLine);
 			this._lineCanvas.addEventListener("mouseup", this._binds.mouseUpLine);
 			this._lineCanvas.addEventListener("contextmenu", this._binds.contextMenu);
 
-			if (this._mouse.wasRightClick) {
+			if (this._flags.wasRightClick) {
 				this._lineCanvas.classList.add("is-dragged");
 			}
 
@@ -662,19 +665,19 @@ function(
 	 */
 	$anonymizer.prototype._mouseMove = function(e) {
 		// mouse cursor
-		if (!this._mouse.wasMove) {
+		if (!this._flags.wasMove) {
 			this._canvas.classList.add("is-dragged");
 		}
 
 		// mouse move flag
-		this._mouse.wasMove = true;
+		this._flags.wasMove = true;
 
 		// mouse move over the preview?
 		var isPreview = this._isPreview(e.offsetX, e.offsetY);
 
-		if (!this._mouse.wasRightClick && !this._mouse.wasImgMove && isPreview) {
+		if (!this._flags.wasRightClick && !this._flags.wasImgMove && isPreview) {
 			// set preview flag
-			this._mouse.wasPreview = true;
+			this._flags.wasPreview = true;
 
 			// image move over the preview
 			this._setPosition(isPreview.xRatio, isPreview.yRatio);
@@ -682,9 +685,9 @@ function(
 			this._alignImgToCanvas();
 			this._redraw();
 		}
-		else if (!this._mouse.wasPreview) {
+		else if (!this._flags.wasPreview) {
 			// image move - flag
-			this._mouse.wasImgMove = true;
+			this._flags.wasImgMove = true;
 
 			// image move
 			this._imgMove(e.offsetX, e.offsetY);
@@ -742,7 +745,7 @@ function(
 		var thresholdTest = false;
 
 		// only it was move
-		if (this._mouse.wasMove) {
+		if (this._flags.wasMove) {
 			// difference towards start click
 			var diffX = this._mouse.startXSave - e.offsetX;
 			var diffY = this._mouse.startYSave - e.offsetY;
@@ -754,7 +757,7 @@ function(
 		}
 
 		// click - there was no move, threshold test, it is disabled for right mouse click
-		if (!this._mouse.wasRightClick && (!this._mouse.wasMove || thresholdTest)) {
+		if (!this._flags.wasRightClick && (!this._flags.wasMove || thresholdTest)) {
 			var isPreview = this._isPreview(e.offsetX, e.offsetY);
 
 			if (isPreview) {
@@ -796,10 +799,10 @@ function(
 	 */
 	$anonymizer.prototype._mouseMoveLine = function(e) {
 		// mouse move
-		this._mouse.wasMove = true;
+		this._flags.wasMove = true;
 
 		// right mouse click
-		if (this._mouse.wasRightClick) {
+		if (this._flags.wasRightClick) {
 			// image move
 			this._imgMove(e.offsetX, e.offsetY);
 
@@ -810,18 +813,18 @@ function(
 		// left mouse click
 		else {
 			var isPreview = this._isPreview(e.offsetX, e.offsetY);
-			var wasPreview = this._mouse.wasPreview;
+			var wasPreview = this._flags.wasPreview;
 
-			if (!this._mouse.wasLine && isPreview) {
-				this._mouse.wasPreview = true;
+			if (!this._flags.wasLine && isPreview) {
+				this._flags.wasPreview = true;
 
 				// move over preview
 				this._setPosition(isPreview.xRatio, isPreview.yRatio);
 				this._alignImgToCanvas();
 				this._redraw();
 			}
-			else if (!this._mouse.wasPreview) {
-				this._mouse.wasLine = true;
+			else if (!this._flags.wasPreview) {
+				this._flags.wasLine = true;
 
 				// line width
 				var lineWidth = Math.round(this._opts.curEntity.value * this._opts.zoom / 100);
@@ -833,7 +836,7 @@ function(
 			}
 
 			// change of state
-			if (!wasPreview && this._mouse.wasPreview) {
+			if (!wasPreview && this._flags.wasPreview) {
 				this._lineCanvas.classList.add("is-dragged");
 			}
 		}
@@ -850,19 +853,19 @@ function(
 	$anonymizer.prototype._mouseUpLine = function(e) {
 		var isPreview = null;
 
-		if (!this._mouse.wasMove) {
+		if (!this._flags.wasMove) {
 			isPreview = this._isPreview(e.offsetX, e.offsetY);
 		}
 
 		// only for left mouse click
-		if (!this._mouse.wasRightClick) {
+		if (!this._flags.wasRightClick) {
 			if (isPreview) {
 				// preview click - click coordinates on the canvas center
 				this._setPosition(isPreview.xRatio, isPreview.yRatio);
 				this._alignImgToCanvas();
 				this._redraw();
 			}
-			else if (this._mouse.wasLine) {
+			else if (this._flags.wasLine) {
 				// create a line
 				var zc = this._opts.zoom / 100;
 				var xc = Math.round(this._x * zc);
@@ -907,7 +910,7 @@ function(
 	 * @member $anonymizer
 	 */
 	$anonymizer.prototype._mouseUpDocument = function(e) {
-		this._mouse.wasLine = false;
+		this._flags.wasLine = false;
 		this._mouseUpLine(e);
 	};